chore: restructure into server/ + shared/ directories
Resolves #4 — separates backend into server/ for future mobile app.

- Move all backend .js files and directories to server/
- Create shared/ directory (with .gitkeep)
- Update ctx.dir to serve static files from repo root
- Update WORKSPACE paths for new directory depth
- Update package.json scripts and .gitignore paths
- Update brief-panel require path
