+++ package.json
@@ -2,10 +2,10 @@
   "name": "@fagemx/task-engine",
   "version": "0.1.0",
   "description": "AI task orchestration engine — board + brief + multi-runtime dispatch",
-  "main": "server.js",
+  "main": "server/server.js",
   "scripts": {
-    "start": "node server.js",
-    "test": "node test-evolution-loop.js"
+    "start": "node server/server.js",
+    "test": "node server/test-evolution-loop.js"
   },
   "keywords": ["ai", "task-engine", "orchestration", "blackboard", "dispatch"],
   "license": "MIT"
