+++ server/docs/architecture.md
