+++ server/docs/plans/EVOLUTION_LAYER/T5_GATE_AND_LESSONS.md
