+++ server/docs/research/mobile-judgment-slot/2026-02-27/00-v0-converged-spec.md
