+++ server/docs/research/mobile-judgment-slot/2026-02-27/A-mobile-habit-model.md
