+++ server/docs/research/mobile-judgment-slot/2026-02-27/B-blackboard-interaction-grammar.md
