+++ server/docs/research/mobile-judgment-slot/2026-02-27/D-failure-modes-guardrails.md
