+++ server/docs/strategic-layer.md
