+++ server/skills/blackboard-basics/SKILL.md
