+++ server/skills/engineer-playbook/SKILL.md
