+++ server/skills/lead-playbook/SKILL.md
