+++ server/smoke-test.js
